+++ app/src/main/java/com/jin/honey/feature/foodsearch/ui/FoodSearchViewModel.kt
@@ -6,6 +6,7 @@ import com.jin.feature.ui.state.SearchState
 import com.jin.domain.repositories.PreferencesRepository
 import com.jin.domain.usecase.SearchMenusUseCase
 import com.jin.domain.food.model.MenuPreview
+import com.jin.domain.search.SearchRepository
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
@@ -13,7 +14,7 @@ import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.launch
 
 class FoodSearchViewModel(
-    private val repository: PreferencesRepository,
+    private val repository: SearchRepository,
     private val searchMenusUseCase: SearchMenusUseCase
 ) : ViewModel() {
     private val _menuSearchState = MutableStateFlow<SearchState<List<MenuPreview>>>(SearchState.Idle)
