+++ domain/src/main/java/com/jin/domain/search/SearchRepository.kt
@@ -1,0 +1,10 @@
+package com.jin.domain.search
+
+import kotlinx.coroutines.flow.Flow
+
+interface SearchRepository {
+    fun flowSearchKeywords(): Flow<List<String>>
+    suspend fun saveSearchKeyword(menuName: String)
+    suspend fun deleteSearchKeyword(menuName: String)
+    suspend fun clearSearchKeyword()
+}
